realm_domains_modal: Migrate modal to dialog_widget.
