+++ frontend_tests/node_tests/settings_org.js
@@ -24,15 +24,6 @@ mock_esm("../../static/js/loading", {
     make_indicator: noop,
     destroy_indicator: noop,
 });
-mock_esm("../../static/js/ui_report", {
-    success(msg, $elem) {
-        $elem.val(msg);
-    },
-
-    error(msg, xhr, $elem) {
-        $elem.val(msg);
-    },
-});
 
 set_global(
     "FormData",
@@ -69,59 +60,9 @@ test("unloaded", () => {
     // sure things don't explode before set_up is called.
 
     settings_org.reset();
-    settings_org.populate_realm_domains();
+    settings_org.populate_realm_domains_label();
     settings_org.populate_auth_methods();
 });
-
-function simulate_realm_domains_table() {
-    const $tr_stub = $.create("realm-tr-stub");
-    $("#realm_domains_table tbody").set_find_results("tr", $tr_stub);
-    $tr_stub.remove = () => {};
-
-    let appended;
-    $("#realm_domains_table tbody").append = (html) => {
-        appended = true;
-        assert.equal(html, "stub-domains-list");
-    };
-
-    return function verify() {
-        assert.ok(appended);
-    };
-}
-
-function test_realms_domain_modal(override, add_realm_domain) {
-    const $info = $(".realm_domains_info");
-
-    $("#add-realm-domain-widget").set_find_results(
-        ".new-realm-domain",
-        $.create("new-realm-domain-stub"),
-    );
-
-    $("#add-realm-domain-widget").set_find_results(
-        ".new-realm-domain-allow-subdomains",
-        $.create("new-realm-domain-allow-subdomains-stub"),
-    );
-
-    let posted;
-    let success_callback;
-    let error_callback;
-    override(channel, "post", (req) => {
-        posted = true;
-        assert.equal(req.url, "/json/realm/domains");
-        success_callback = req.success;
-        error_callback = req.error;
-    });
-
-    add_realm_domain();
-
-    assert.ok(posted);
-
-    success_callback();
-    assert.equal($info.val(), "translated HTML: Added successfully!");
-
-    error_callback({});
-    assert.equal($info.val(), "translated HTML: Failed");
-}
 
 function createSaveButtons(subsection) {
     const $stub_save_button_header = $(`#org-${CSS.escape(subsection)}`);
@@ -346,56 +287,6 @@ function test_upload_realm_icon(override, upload_realm_logo_or_icon) {
 
     upload_realm_logo_or_icon(file_input, null, true);
     assert.ok(posted);
-}
-
-function test_change_allow_subdomains(change_allow_subdomains) {
-    const ev = {
-        stopPropagation: noop,
-    };
-
-    const $info = $(".realm_domains_info");
-    $info.fadeOut = noop;
-    const domain = "example.com";
-    let allow = true;
-
-    let success_callback;
-    let error_callback;
-    channel.patch = (req) => {
-        assert.equal(req.url, "/json/realm/domains/example.com");
-        assert.equal(req.data.allow_subdomains, JSON.stringify(allow));
-        success_callback = req.success;
-        error_callback = req.error;
-    };
-
-    const $domain_obj = $.create("domain object");
-    $domain_obj.text(domain);
-
-    const $elem_obj = $.create("<elem html>");
-    const $parents_obj = $.create("parents object");
-
-    $elem_obj.set_parents_result("tr", $parents_obj);
-    $parents_obj.set_find_results(".domain", $domain_obj);
-    $elem_obj.prop("checked", allow);
-
-    change_allow_subdomains.call($elem_obj, ev);
-
-    success_callback();
-    assert.equal(
-        $info.val(),
-        "translated HTML: Update successful: Subdomains allowed for example.com",
-    );
-
-    error_callback({});
-    assert.equal($info.val(), "translated HTML: Failed");
-
-    allow = false;
-    $elem_obj.prop("checked", allow);
-    change_allow_subdomains.call($elem_obj, ev);
-    success_callback();
-    assert.equal(
-        $info.val(),
-        "translated HTML: Update successful: Subdomains no longer allowed for example.com",
-    );
 }
 
 function test_extract_property_name() {
@@ -637,10 +528,7 @@ function test_discard_changes_button(discard_changes) {
     settings_org.__Rewire__("change_save_button_state", stubbed_function);
 }
 
-test("set_up", ({override, override_rewire, mock_template}) => {
-    mock_template("settings/admin_realm_domains_list.hbs", false, () => "stub-domains-list");
-
-    const verify_realm_domains = simulate_realm_domains_table();
+test("set_up", ({override, override_rewire}) => {
     page_params.realm_available_video_chat_providers = {
         jitsi_meet: {
             id: 1,
@@ -690,9 +578,6 @@ test("set_up", ({override, override_rewire, mock_template}) => {
     override_rewire(settings_org, "maybe_disable_widgets", noop);
     settings_org.set_up();
 
-    verify_realm_domains();
-
-    test_realms_domain_modal(override, () => $("#submit-add-realm-domain").trigger("click"));
     test_submit_settings_form(
         override,
         $(".admin-realm-form").get_on_handler(
@@ -701,9 +586,6 @@ test("set_up", ({override, override_rewire, mock_template}) => {
         ),
     );
     test_upload_realm_icon(override, upload_realm_logo_or_icon);
-    test_change_allow_subdomains(
-        $("#realm_domains_table").get_on_handler("change", ".allow-subdomains"),
-    );
     test_extract_property_name();
     test_change_save_button_state();
     test_sync_realm_settings();
