+++ static/templates/settings/admin_settings_modals.hbs
@@ -1,4 +1,2 @@
-{{> realm_domains_modal }}
-
 <div id="user-info-form-modal-container"></div>
 <div id="linkifier-edit-form-modal-container"></div>
