+++ static/templates/settings/organization_permissions_admin.hbs
@@ -32,7 +32,7 @@
                 <div class="dependent-block">
                     <p id="allowed_domains_label" class="inline-block"></p>
                     {{#if is_admin }}
-                    <a data-toggle="modal" href="#realm_domains_modal">{{t "[Configure]" }}</a>
+                    <a id="show_realm_domains_modal" role="button">{{t "[Configure]" }}</a>
                     {{/if}}
                 </div>
                 <div class="input-group">
