+++ static/templates/settings/realm_domains_modal.hbs
@@ -1,34 +1,22 @@
-<div id="realm_domains_modal" class="modal modal-bg hide" tabindex="-1" role="dialog" aria-labelledby="realm_domains_modal_label" aria-hidden="true">
-    <div class="modal-header">
-        <button type="button" class="close" data-dismiss="modal" aria-label="{{t 'Close' }}"><span aria-hidden="true">&times;</span></button>
-        <h3 id="realm_domains_modal_label">{{t "Allowed domains" }}</h3>
-    </div>
-    <div class="modal-body">
-        <table class="table table-condensed table-stripped" id="realm_domains_table">
-            <thead>
-                <th>{{t "Domain" }}</th>
-                <th>{{t "Allow subdomains"}}</th>
-                <th>{{t "Action" }}</th>
-            </thead>
-            <tbody>
-            </tbody>
-            <tfoot>
-                <tr id="add-realm-domain-widget">
-                    <td><input type="text" class="new-realm-domain" placeholder="acme.com" /></td>
-                    <td>
-                        <label class="checkbox">
-                            <input type="checkbox" class="new-realm-domain-allow-subdomains" />
-                            <span></span>
-                        </label>
-                    </td>
-                    <td><button type="button" class="button sea-green small rounded" id="submit-add-realm-domain">{{t "Add" }}</button></td>
-                </tr>
-            </tfoot>
-        </table>
-
-        <div class="alert realm_domains_info"></div>
-    </div>
-    <div class="modal-footer centered-footer">
-        <button class="new-style button rounded" data-dismiss="modal">{{t "Close" }}</button>
-    </div>
-</div>
+<table class="table table-condensed table-stripped new-style" id="realm_domains_table">
+    <thead>
+        <th>{{t "Domain" }}</th>
+        <th>{{t "Allow subdomains"}}</th>
+        <th>{{t "Action" }}</th>
+    </thead>
+    <tbody>
+    </tbody>
+    <tfoot>
+        <tr id="add-realm-domain-widget">
+            <td><input type="text" class="new-realm-domain" placeholder="acme.com" /></td>
+            <td>
+                <label class="checkbox">
+                    <input type="checkbox" class="new-realm-domain-allow-subdomains" />
+                    <span></span>
+                </label>
+            </td>
+            <td><button type="button" class="button sea-green small rounded" id="submit-add-realm-domain">{{t "Add" }}</button></td>
+        </tr>
+    </tfoot>
+</table>
+<div class="alert realm_domains_info"></div>
